bitfield.h: add continuing versions of read funcs
The current version of all the bitfield read functions
start the accumulator either at 0 or -1 depending on the
the signedness of the value to decode. This patch
extends the read functions, creating readx versions,
to initialize the accumulator with the last returned
value instead (read from _vptr). This allows reading
a single bitfield from multiple, independent buffers,
by first calling the function with _cont set to 0, then
calling it again with _cont set to 1.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>