Initial implementation of the debuginfo API
This patchset adds so_info and durin, which are used in debug info or
callsite analysis. The developer is expected to use the SO info API
documented in `include/babeltrace/ctf-text/so-info.h`. Durin is only
used internally by so_info as a wrapper around libdwarf.

There is a hard dependency on libelf and libdwarf, currently not yet
configurable via the configure script.

Signed-off-by: Antoine Busque <abusque@efficios.com>