Add env wrapper to facilitate importing the in-tree python bindings
The test_python_bt2 script sets up a special environment for running the
Python bindings tests.  Crucially, it sets up PYTHONPATH so we are able
to "import bt2" from a Python interpreter.  It can be difficult to
reproduce this environment exactly in an interactive shell, or when
trying to run a single test.

This patch introduces test_python_bt2_env, which can be used to run a
command in the same environment.  This command can be anything from a
single test run to a sub-shell.

Also, we expect that many Babeltrace tests will test some non-Python
features using Python, for convenience.  These tests will also need to
run in the right environment to be able to import the bt2 Python
package.  Therefore, I think it makes sense to put it under tests/utils.

Once the bindings and their tests work better, we should make
the test_python_bt2 script use test_python_bt2_env, to avoid the duplication, but
for now I prefer to leave it as-is.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Change-Id: I4e1655614756d638ef6bb9a4426719d4b754aa4b