tests/bindings/python: Mark all tests as skipped
This patch makes us skip all the Python bindings tests.  The goal is for
these commands to run with success (even though they doesn't actually test
anything):

    $ tests/bindings/python/bt2/test_python_bt2
    $ tests/bindings/python/babeltrace/test_python_babeltrace

The strategy will be to gradually fix the bindings and their tests, and
gradually remove those skip annotations.  At any point, we should be
able to run the testsuite with success, ensuring that we don't regress.

After this patch, "make check" is clean for me, with python bindings
enabled (--enable-python-bindings).

I am not sure what is still relevant in the python/babeltrace
directory, I suppose the ctf writer stuff still is.  I think we can sort
this out later.

Change-Id: I1169c255c0c7c29fa5de75a358c02c4cd57aae20
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>