Fix: update_field_class_in_ir(): do not have special int fields in trace IR
The logic in update_field_class_in_ir() for the integer field class case
was wrong. This patch fixes this, so that we only translate the CTF IR
integer field class to a trace IR integer field class if another field
class depends on it (dynamic array, variant) or if it does not update a
clock AND it does not have a special meaning.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>