ctf: Add dash to stream name suffixes
We add numerical suffixes to stream names, if needed, to make them
unique. For trace directories, we do something similar and decided to
format these suffixes with a dash to avoid confusion if the original
name ends with a number.  This patch does the same thing but for stream
names.

Change-Id: I3cb0bca457575414d9963fe6d4ad5cc700dba38b
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>