src/logging/log.h: add BT_LOG_WRITE_CUR_LVL(), BT_LOG_WRITE_ERRNO_CUR_LVL()
BT_LOG_WRITE_CUR_LVL() is like BT_LOG_WRITE(), but it compares the
given log level to a current log level parameter instead of
`_BT_LOG_OUTPUT_LEVEL`.

BT_LOG_WRITE_ERRNO_CUR_LVL() is the same thing, but for
BT_LOG_WRITE_ERRNO().

The goal of this is to be able to control which variable contains the
current log level at the call site to achieve per-component log levels.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I670664a7cb465954c2f0f8b3551ea49fd1465e32