lib: pass log level to bt_query_executor_query()
This new parameter indicates the log level to use during the query
execution.

The project's component classes are changed to receive this log level,
but they do not use it. This work is reserved for a future patch.

The CLI passes its default log level (main `--log-level` option) to
bt_query_executor_query().

Python bindings are adapted accordingly.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I929d545a1df2f4d6d083f6c536f6fcf26aa972bd