`ctf` plugin: metadata: use local log level
As of this patch, the users of the common metadata functions and objects
set the new log level parameters to their own (global) log level.

The goal is to eventually have a per-component log level for the `ctf`
plugin's component classes.

Unfortunately, `YYFPRINTF` and `YY_FATAL_ERROR` are pretty limited in
terms of available context. I'm leaving the
`BABELTRACE_PLUGIN_CTF_METADATA_LOG_LEVEL` environment variable to set
the logging level of the parser and lexer modules.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I34651e6d0c14b5acc9b4198681c2aea4c003dc7a