lib: add boolean field class and field types
This patch adds the boolean field class and field to the available
library types. They are analogous to the boolean value object which
already exists.

CTF 2 will very likely have such a type, so Babeltrace 2.0 will be
ready.

This patch does not adapt existing plugins and the Python bindings to
use and wrap the new boolean field class and field; this work is
reserved for subsequent patches.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I90874f214757ad67d5eecaad5bd8297b8610c001